requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.5
django-location-field==2.1.0
django-shortcuts==1.6
djangorestframework==3.12.4
drf-spectacular==0.17.2
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
jsonschema==3.2.0
Markdown==3.3.4
psycopg==3.0.14
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.9
whitenoise==6.1.0
